Resolve circular import by moving get_config_dir
